+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2022-05-28
 ### Added
 - DotNet template for toolRestore and restore in one step named restoreWithTools
 - DotNet template for basic packing
@@ -38,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-10-27
 Initial release
 
-[Unreleased]: https://github.com/NicoVIII/RunHelpers/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/NicoVIII/RunHelpers/compare/v0.5.0...HEAD
+[0.4.1]: https://github.com/NicoVIII/RunHelpers/compare/v0.4.1..v0.5.0
 [0.4.1]: https://github.com/NicoVIII/RunHelpers/compare/v0.4.0..v0.4.1
 [0.4.0]: https://github.com/NicoVIII/RunHelpers/compare/v0.3.0..v0.4.0
 [0.3.0]: https://github.com/NicoVIII/RunHelpers/compare/v0.2.0..v0.3.0